requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.7.0
attrs==24.2.0
audioop-lts==0.2.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comtypes==1.4.8
distro==1.9.0
frozenlist==1.5.0
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
MouseInfo==0.1.3
multidict==6.1.0
openai==0.28.0
propcache==0.2.1
psutil==6.1.0
PyAutoGUI==0.9.54
pycaw==20240210
pydantic==2.10.3
pydantic_core==2.27.1
pydub==0.25.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.16.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dotenv==1.0.1
pytweening==1.2.0
requests==2.32.3
sniffio==1.3.1
SpeechRecognition==3.12.0
standard-aifc==3.13.0
standard-chunk==3.13.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.18.3
